requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
certifi==2018.11.29
cffi==1.11.5
cryptography==2.5
future==0.17.1
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.7.0
six==1.12.0
soupsieve==1.7.3
